Fix #1 by checking if pip already exists
This is an (preliminary) fix for the issues reported in #1, further improvements may be necessary.
Blender 2.82a for Windows is distributed without ensurepip, but with pip. Importing ensurepip
therefore fails. The patch checks if pip is already installed and only imports ensurepip if it
isn't. Since either pip is already installed or ensurepip is available, this should allow the
add-on to work on all current Blender versions.